pyproject.toml
[project]
name = "customer-service-bot"
version = "0.1.0"
description = "A simple website chatbot powered by FastAPI and Gemini."
authors = [
    {name = "Nevil-ing", email = "nevilpurpp12@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "uvicorn[standard]>=0.29.0,<0.30.0",
    "fastapi>=0.111.0,<0.112.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "requests>=2.32.0,<3.0.0",
    "scrapy>=2.11.2,<3.0.0",
    "google-generativeai>=0.5.4,<0.6.0",
    "jinja2>=3.1.4,<4.0.0",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"